requirements.txt
anthropic
pyyaml
python-dotenv
requests
yfinance
pandas
ta
schedule
fredapi
matplotlib
